Further refactoring. Moved outputLocs and associated functions to inside ShuffleMapStage.
